Update generic_strings.csv
Hello. I made some changes to the Spanish translations. Mostly minor stuff like accentuation. I also fixed the translations for "male" and "female." When referring to a person's gender, it is more appropriate to use "macho" and "hembra" for male and female respectively. The use of "sexo" to describe an individual's gender is also more desirable. Also fixed the "Are you sure you want to quit?" prompt as well as the test_character_follow and wait dialogues. I double checked to see if I am using correct grammar but feel free to have someone check on it. I hope this helps and good luck with your project. I look forward to seeing it evolve into something awesome! :)